+++ lib/src/app/examples/contacts/src/home/view/view.dart
@@ -32,3 +32,9 @@ class MyHome extends StatefulWidget {
   @override
   State createState() => App.useMaterial ? MyAndroid() : MyiOS();
 }
+
+
+
+
+
+