+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+ May 06, 2020
+- ContactListState() : super(Contact()) {
+- model.dart import file reformatted.
+
 ## 0.1.0
  May 02, 2020
 - Initial Commit